+++ rust_compute/src/queen_mask.rs
@@ -10,7 +10,7 @@ pub fn queen_moves(chess_board: &mut ChessBoard, player_index: usize, opp_index:
     let mut queen_bitboard = chess_board.queens[player_index];
 
     while queen_bitboard != 0 {
-        let queen = queen_bitboard.trailing_zeros() as usize;
+        let queen = queen_bitboard.trailing_zeros() as u8;
 
         let rook_attack_paths = rook_attack_paths(queen, chess_board.occupied);
         let bishop_attack_paths = bishop_attack_paths(queen, chess_board.occupied);
@@ -20,7 +20,7 @@ pub fn queen_moves(chess_board: &mut ChessBoard, player_index: usize, opp_index:
         let mut queen_captures = total_attacks & chess_board.all_pieces[opp_index];
 
         while queen_moves != 0 {
-            let target = queen_moves.trailing_zeros() as usize;
+            let target = queen_moves.trailing_zeros() as u8;
             moves.push(ForwardMove { 
                 start_sq: queen, end_sq: target, move_type: MoveFlag::MOVE, pv_score: 1000 
             });
@@ -28,7 +28,7 @@ pub fn queen_moves(chess_board: &mut ChessBoard, player_index: usize, opp_index:
         }
 
         while queen_captures != 0 {
-            let target = queen_captures.trailing_zeros() as usize;
+            let target = queen_captures.trailing_zeros() as u8;
 
             let captured_piece_val = piece_value(chess_board.mailbox_piece(target));
             let pv_score = 100 - (captured_piece_val * 10) + 5;
